requirements/base.txt
django-registration==1.0
django-reversion==1.7.1
Django==1.5.4
